+++ src/hooks/index.ts
@@ -5,3 +5,4 @@ export { useSessionStorage } from './useSessionStorage';
 export { useCookie } from './useCookie';
 export { useClickOutside } from './useClickOutside';
 export { usePrevious } from './usePrevious';
+export { useToggle } from './useToggle';